+++ cli/cli/commands/config/init/init.go
@@ -10,6 +10,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/prompt_displayer"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config"
 	"github.com/kurtosis-tech/stacktrace"
+	"github.com/mattn/go-isatty"
 	"github.com/sirupsen/logrus"
 	"os"
 )
@@ -87,7 +88,7 @@ func run(ctx context.Context, flags *flags.ParsedFlags, args *args.ParsedArgs) e
 	if doesKurtosisConfigAlreadyExists && !shouldForceInitConfig {
 		// Check if we're actually running in interactive mode (i.e. STDOUT is a terminal) before displaying
 		//  the interactive prompt
-		if fileInfo, _ := os.Stdout.Stat(); (fileInfo.Mode() & os.ModeCharDevice) != 0 {
+		if !(isatty.IsTerminal(os.Stdout.Fd()) || isatty.IsCygwinTerminal(os.Stdout.Fd())) {
 			return stacktrace.NewError(
 				"The Kurtosis config already exists and the '%v' flags wasn't specified so this is where we'd normally ask for " +
 					"a confirmation to overwrite the config, except STDOUT isn't a terminal (indicating that this is probably " +